python/requirements/integration.txt
# This file was manually created for integration test dependencies.
# To update, run:
#
#    pip-compile requirements/integration.in -o requirements/integration.txt
#
-r base.txt
google-cloud-bigquery==3.31.0
    # via -r requirements/integration.in
google-cloud-core==2.5.0
    # via google-cloud-bigquery
google-cloud-dataproc==5.25.0
    # via -r requirements/integration.in
requests==2.32.5
    # via -r requirements/integration.in
